+++ README.md
@@ -52,3 +52,18 @@ Given orderId 를 이용해 상품을 받아올 수 있는 상태이다.
 When orderId 를 이용해 주문 상세 정보를 받아온다.
 Then 주문 상세 정보를 띄운다.
 ```
+
+## 주문 목록 페이지
+
+- [ ] 화면에 주문 목록이 나타난다.
+```gherkin
+Given 
+When 주문 목록을 받아온다.
+Then 주문 목록을 띄운다.
+```
+- [ ] 주문 목록을 클릭하면 주문 상세 화면으로 넘어간다.
+```gherkin
+Given 주문 목록이 불러와진 상태이다.
+When 주문 목록을 클릭한다.
+Then 해당 주문 상세 화면으로 넘어간다.
+```
