+++ app/src/main/AndroidManifest.xml
@@ -15,10 +15,10 @@
         android:usesCleartextTraffic="true"
         tools:targetApi="31">
         <activity
-            android:name=".feature.order.OrderActivity"
+            android:name=".feature.order.order.OrderActivity"
             android:exported="false" />
         <activity
-            android:name=".feature.orderdetail.OrderDetailActivity"
+            android:name=".feature.order.detail.OrderDetailActivity"
             android:exported="false" />
         <activity
             android:name=".feature.serversetting.ServerSettingActivity"
