+++ app/src/main/java/woowacourse/shopping/feature/cart/CartActivity.kt
@@ -12,7 +12,7 @@ import woowacourse.shopping.R
 import woowacourse.shopping.data.repository.local.CartRepositoryImpl
 import woowacourse.shopping.data.service.cart.CartRemoteService
 import woowacourse.shopping.databinding.ActivityCartBinding
-import woowacourse.shopping.feature.order.OrderActivity
+import woowacourse.shopping.feature.order.order.OrderActivity
 import woowacourse.shopping.util.toMoneyFormat
 
 class CartActivity : AppCompatActivity(), CartContract.View {
