+++ app/src/main/java/woowacourse/shopping/feature/order/detail/OrderDetailPresenter.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.feature.orderdetail
+package woowacourse.shopping.feature.order.detail
 
 import com.example.domain.repository.OrderRepository
 import woowacourse.shopping.mapper.toPresentation
